+++ RedProtect-Spigot/src/main/java/br/net/fabiozumbi12/RedProtect/Bukkit/Region.java
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2012-2024 - @FabioZumbi12
- * Last Modified: 24/06/2024 18:53
+ * Last Modified: 26/06/2024 16:48
  *
  * This class is provided 'as-is', without any express or implied warranty. In no event will the authors be held liable for any
  *  damages arising from the use of this class.
@@ -70,7 +70,7 @@ public class Region extends CoreRegion {
      * @param candel    Can delete?
      * @param canPurge  Can be purged?
      */
-    public Region(String name, Set<PlayerRegion> admins, Set<PlayerRegion> members, Set<PlayerRegion> leaders, Location minLoc, Location maxLoc, Map<String, Object> flags, String wMessage, int prior, String worldName, String date, long value, Location tppoint, boolean candel, boolean canPurge) {
+    public Region(String name, Set<PlayerRegion> admins, Set<PlayerRegion> members, Set<PlayerRegion> leaders, Location minLoc, Location maxLoc, Map<String, Object> flags, String wMessage, int prior, String worldName, String date, double value, Location tppoint, boolean candel, boolean canPurge) {
         super(name, admins, members, leaders, new int[]{minLoc.getBlockX(), minLoc.getBlockY(), minLoc.getBlockZ()}, new int[]{maxLoc.getBlockX(), maxLoc.getBlockY(), maxLoc.getBlockZ()}, flags, wMessage, prior, worldName, date, value, tppoint == null ? null : new int[]{tppoint.getBlockX(), tppoint.getBlockY(), tppoint.getBlockZ()}, tppoint == null ? null : new float[]{tppoint.getPitch(), tppoint.getYaw()}, candel, canPurge);
         checkParticle();
     }
@@ -98,7 +98,7 @@ public class Region extends CoreRegion {
      * @param candel    Can delete?
      * @param canPurge  Can br purged?
      */
-    public Region(String name, Set<PlayerRegion> admins, Set<PlayerRegion> members, Set<PlayerRegion> leaders, int maxMbrX, int minMbrX, int maxMbrZ, int minMbrZ, int minY, int maxY, Map<String, Object> flags, String wMessage, int prior, String worldName, String date, long value, Location tppoint, boolean candel, boolean canPurge) {
+    public Region(String name, Set<PlayerRegion> admins, Set<PlayerRegion> members, Set<PlayerRegion> leaders, int maxMbrX, int minMbrX, int maxMbrZ, int minMbrZ, int minY, int maxY, Map<String, Object> flags, String wMessage, int prior, String worldName, String date, double value, Location tppoint, boolean candel, boolean canPurge) {
         super(name, admins, members, leaders, maxMbrX, minMbrX, maxMbrZ, minMbrZ, minY, maxY, flags, wMessage, prior, worldName, date, value, tppoint == null ? null : new int[]{tppoint.getBlockX(), tppoint.getBlockY(), tppoint.getBlockZ()}, tppoint == null ? null : new float[]{tppoint.getPitch(), tppoint.getYaw()}, candel, canPurge);
         checkParticle();
     }
@@ -124,7 +124,7 @@ public class Region extends CoreRegion {
      * @param candel    Can delete?
      * @param canPurge  Can be purged?
      */
-    public Region(String name, Set<PlayerRegion> admins, Set<PlayerRegion> members, Set<PlayerRegion> leaders, int[] x, int[] z, int miny, int maxy, int prior, String worldName, String date, Map<String, Object> flags, String welcome, long value, Location tppoint, boolean candel, boolean canPurge) {
+    public Region(String name, Set<PlayerRegion> admins, Set<PlayerRegion> members, Set<PlayerRegion> leaders, int[] x, int[] z, int miny, int maxy, int prior, String worldName, String date, Map<String, Object> flags, String welcome, double value, Location tppoint, boolean candel, boolean canPurge) {
         super(name, admins, members, leaders, x, z, miny, maxy, prior, worldName, date, flags, welcome, value, tppoint == null ? null : new int[]{tppoint.getBlockX(), tppoint.getBlockY(), tppoint.getBlockZ()}, tppoint == null ? null : new float[]{tppoint.getPitch(), tppoint.getYaw()}, candel, canPurge);
         checkParticle();
     }
@@ -1187,7 +1187,7 @@ public class Region extends CoreRegion {
     }
     //--------------------------------------------------------------//
 
-    public void setValue(long value) {
+    public void setValue(double value) {
         setToSave(true);
         RedProtect.get().getRegionManager().updateLiveRegion(this, "value", String.valueOf(value));
         this.value = value;
